given insertion rate, calculate theoretical distribution under condition 1 and 2

given insertion rate, calculate theoretical distribution under condition 1 and 2

Usage

.insertion_to_c12(
  insertion_rates = (1:2000) * 0.01,
  peak_lengths = 4:20 * 50,
  pad_length = 0,
  dirs = c(-1, 1),
  count_type = "fragment",
  min_frag_length = 25,
  max_frag_length = 600,
  cap_insertion = 20
)

Arguments

insertion_rates

A vector of insertion rates, note that the insertion rate is per 1000 base, so the default range is from 0.001 to 20

peak_lengths

A vector of peak length ranges, default is from 200 to 1000

pad_length

how many bp outside of peaks can insertions happen, default = 0

dirs

directions. default (-1, 1)

count_type

what type of count should we estimate, default 'fragment'

min_frag_length

minimum of fragment length, default 25

max_frag_length

maximum of fragment length, default 600

cap_insertion

maximum number of insertions that can simultaneously happen within one peak region

Value

A insertion rate by peak length matrix, and each element of the matrix corresponds to the number of expected fragments #' @noRd